+++ test/state/auth_notifier_test.dart
@@ -83,6 +83,13 @@ class FakeUserRepository implements UserRepository {
 
   @override
   Future<Either<Failure, void>> deleteKey(int id) => throw UnimplementedError();
+
+  @override
+  Future<Either<Failure, List<Activity>>> getUserActivities(
+    String username, {
+    int? page,
+    int? limit,
+  }) => throw UnimplementedError();
 }
 
 class FakeMiscRepository implements MiscRepository {